react-dom-route added, files moved to components
