+++ package.json
@@ -8,6 +8,7 @@
     "react": "^16.2.0",
     "react-dom": "^16.2.0",
     "react-particles-js": "^2.1.0",
+    "react-router-dom": "^4.2.2",
     "react-scripts": "1.1.0",
     "react-tap-event-plugin": "^3.0.2"
   },
